Real Estate Backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
GDAL @ file:///C:/Users/PC/Documents/Workspace/Django/Real%20Estate/Real%20Estate%20Backend/GDAL-3.4.3-cp311-cp311-win_amd64.whl
psycopg2==2.9.5
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
